Allow editing history in the console.
The idea is to be similar to how readline (e.g., the terminal in bash)
works: when you're back in the history, if you start typing, you don't
lose your place in the history.  And if you move up/down after typing,
your changes are saved for when you come back.  (If you don't move,
though, the entry in your history doesn't change.)

This seems incompatible with autocompletion.  Fortunately, there's
already an option that's incompatible with autocompletion:
!mHighlightHistory.  So I can gate the code for this behind that
option.